Move dg inspect to dg utils inspect-component-type and remove it from the docs
Summary:
https://linear.app/dagster-labs/issue/BUILD-735/default-dg-inspect-to-human-readable-output - keep the functionality while keeping it from confusing people
