+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/1-help.txt
@@ -39,7 +39,6 @@ Usage: dg [OPTIONS] COMMAND [ARGS]...
 │ docs       Commands for generating docs from your Dagster code.                        │
 │ init       Initialize a new Dagster workspace and a first project within that          │
 │            workspace.                                                                  │
-│ inspect    Commands for inspecting Dagster entities.                                   │
 │ list       Commands for listing Dagster entities.                                      │
 │ scaffold   Commands for scaffolding Dagster code.                                      │
 │ utils      Assorted utility commands.                                                  │
