--- examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/19-dg-component-type-info.txt
@@ -1,36 +1,0 @@
-dg inspect component-type 'dbt_project@dagster_components'
-
-dbt_project@dagster_components
-
-Description:
-
-Expose a DBT project to Dagster as a set of assets.
-
-Scaffold params schema:
-
-{
-    "properties": {
-        "init": {
-            "default": false,
-            "title": "Init",
-            "type": "boolean"
-        },
-        "project_path": {
-            "anyOf": [
-                {
-                    "type": "string"
-                },
-                {
-                    "type": "null"
-                }
-            ],
-            "default": null,
-            "title": "Project Path"
-        }
-    },
-    "title": "DbtScaffoldSchema",
-    "type": "object"
-}
-
-Component schema:
-...
