+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/19-dg-scaffold-jdbt.txt
