+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/20-component-jdbt.yaml
