+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/22-dg-component-check-error.txt
