+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/23-project-jdbt.yaml
@@ -1,0 +1,7 @@
+type: dbt_project@dagster_components
+
+attributes:
+  dbt:
+    project_dir: ../../../dbt/jdbt
+  asset_attributes:
+    key: "target/main/{{ node.name }}"
