+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/24-dg-component-check.txt
