+++ examples/docs_beta_snippets/docs_beta_snippets/guides/components/index/25-duckdb-select-orders.txt
